pyproject.toml
[project]
name = "board-anon-bot"
version = "0.2.0"
description = "Anonymous multi-board Telegram bot with in-bot admin management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiogram",
  "sqlmodel",
  "sqlalchemy",
  "pydantic-settings",
  "python-slugify",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "httpx",
  "ruff",
  "ty==0.0.18",
]

[project.scripts]
board-anon-bot = "app.main:run"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ty.environment]
python-version = "3.13"
python = ".venv"

[tool.ty.src]
include = ["app", "tests", "main.py"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
